+++ App/Views/Reviews/article.view.php
@@ -15,7 +15,7 @@ $review = $data;
         <p class="review_text"><?=$review->getParagraph3()?></p>
         <img src="<?=$review->getImageSrc()?>" alt="<?=$review->getImageAlt()?>" class="review_img">
         <p class="review_text"><?=$review->getParagraph4()?></p>
-        <?php if ($auth->isLogged()) { ?>
+        <?php if ($auth->isLogged() && $auth->getLoggedUserId() == 10) { ?>
             <div class="articlebtns">
                 <a href="?c=reviews&a=editArticleForm&id=<?=$review->getId()?>" class="btn btn-primary articlebtn">E</a>
                 <a href="?c=reviews&a=deleteArticle&id=<?=$review->getId()?>" onclick="return confirm('Recenzia bude zmazaná')" class="btn btn-danger articlebtn">X</a>
